day7

+++ day1/solution.c
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #include "parse.h"
 #include "heapsort.h"
 #include "abs.h"
@@ -23,7 +25,7 @@ static int parse_input(const char * input, int length, struct list * list)
   return i;
 }
 
-int day1_part1(const char * input, int length)
+int64_t day1_part1(const char * input, int length)
 {
   struct list list;
   int list_length = parse_input(input, length, &list);
@@ -40,7 +42,7 @@ int day1_part1(const char * input, int length)
   return sum;
 }
 
-int day1_part2(const char * input, int length)
+int64_t day1_part2(const char * input, int length)
 {
   struct list list;
   int list_length = parse_input(input, length, &list);
